add user performance per action report

+++ Data/Data/ArtUserPerformancePerActionUser.cs
@@ -9,14 +9,14 @@ namespace Data.Data
     public class ArtUserPerformancePerActionUser
     {
         public string ACTION_USER { get; set; }
-        public decimal? TOTAL_NUMBER_OF_CASES { get; set; }
-        public decimal? DURATIONS_IN_SECONDS { get; set; }
-        public decimal? AVG_DURATIONS_IN_SECONDS { get; set; }
-        public decimal? DURATIONS_IN_MINUTES { get; set; }
-        public decimal? AVG_DURATIONS_IN_MINUTES { get; set; }
-        public decimal? DURATIONS_IN_HOURS { get; set; }
-        public decimal? AVG_DURATIONS_IN_HOURS { get; set; }
-        public decimal? DURATIONS_IN_DAYS { get; set; }
-        public decimal? AVG_DURATIONS_IN_DAYS { get; set; }
+        public int? TOTAL_NUMBER_OF_CASES { get; set; }
+        public int? DURATIONS_IN_SECONDS { get; set; }
+        public int? AVG_DURATIONS_IN_SECONDS { get; set; }
+        public int? DURATIONS_IN_MINUTES { get; set; }
+        public int? AVG_DURATIONS_IN_MINUTES { get; set; }
+        public int? DURATIONS_IN_HOURS { get; set; }
+        public int? AVG_DURATIONS_IN_HOURS { get; set; }
+        public int? DURATIONS_IN_DAYS { get; set; }
+        public int? AVG_DURATIONS_IN_DAYS { get; set; }
     }
 }
